+++ packages/frontend/src/components/screens/MainScreen/MainScreen.tsx
@@ -198,7 +198,7 @@ export default function MainScreen({ accountId }: Props) {
                   <Icon icon='arrow-left' className='backButtonIcon'></Icon>
                 </Button>
               </span>
-              <div className='archived-chats-title' data-no-drag-region>
+              <div className={styles.archivedChatsTitle} data-no-drag-region>
                 {tx('chat_archived_chats_title')}
               </div>
             </>