+++ packages/target-tauri/src-tauri/src/help_window.rs
@@ -1,12 +1,18 @@
-use log::warn;
-use tauri::{Manager, WebviewWindow};
+use log::{error, warn};
+use tauri::{Manager, State};
 
-use crate::settings::apply_content_protection;
+use crate::{
+    menus::{float_on_top::set_float_on_top_based_on_main_window, help_menu::create_help_menu},
+    settings::{apply_content_protection, apply_zoom_factor_help_window},
+    state::menu_manager::MenuManger,
+};
 
 #[derive(Debug, thiserror::Error)]
 pub(crate) enum Error {
     #[error(transparent)]
     Tauri(#[from] tauri::Error),
+    #[error("MenuCreation {0}")]
+    MenuCreation(String),
 }
 
 impl serde::Serialize for Error {
@@ -19,8 +25,9 @@ impl serde::Serialize for Error {
 }
 
 #[tauri::command]
-pub(crate) fn open_help_window(
+pub(crate) async fn open_help_window(
     app: tauri::AppHandle,
+    menu_manager: State<MenuManger>,
     locale: &str,
     anchor: Option<&str>,
 ) -> Result<(), Error> {
@@ -29,10 +36,19 @@ pub(crate) fn open_help_window(
         url = "help/en/help.html".to_owned();
         warn!("Did not find help file for language {locale}, falling back to english");
     }
-
+    if let Some(anchor) = anchor {
+        url.push('#');
+        url.push_str(anchor);
+    }
     let app_url = tauri::WebviewUrl::App(url.into());
 
+
     let help_window: WebviewWindow = if let Some(help_window) = app.get_webview_window("help") {
+        // TODO theoretically the URL here could still be
+        // about:blank if it has not loaded yet.
+        let mut url = help_window.url()?;
+        url.set_fragment(anchor);
+        help_window.navigate(url)?;
         help_window
     } else {
         tauri::WebviewWindowBuilder::new(&app, "help", app_url.clone()).build()?
@@ -43,13 +59,20 @@ pub(crate) fn open_help_window(
         help_window.show()?;
     }
 
-    let mut url = help_window.url()?;
-    url.set_fragment(anchor);
-    help_window.navigate(url)?;
-
     help_window.set_title("Delta Chat Tauri - Help")?; // TODO: translate help in the title.
 
     let _ = apply_content_protection(&app);
+    let _ = apply_zoom_factor_help_window(&app);
+
+    let _ = set_float_on_top_based_on_main_window(&help_window);
+
+    let help_window_clone = help_window.clone();
+    tauri::async_runtime::block_on(menu_manager.register_window(
+        &app,
+        &help_window,
+        Box::new(move |app| create_help_menu(app, &help_window_clone)),
+    ))
+    .map_err(|err| Error::MenuCreation(err.to_string()))?;
 
     Ok(())
 }