+++ packages/target-tauri/src-tauri/src/state/app.rs
@@ -8,7 +8,10 @@ use anyhow::{bail, Context};
 use log::info;
 use tauri::{AppHandle, Manager};
 
-use crate::temp_file::{clear_tmp_folder, create_tmp_folder};
+use crate::{
+    i18n::get_all_languages,
+    temp_file::{clear_tmp_folder, create_tmp_folder},
+};
 
 #[derive(Default)]
 pub(crate) struct InnerAppState {
@@ -20,6 +23,11 @@ pub(crate) struct AppState {
     pub(crate) inner: Arc<Mutex<InnerAppState>>,
     pub(crate) startup_timestamp: SystemTime,
     pub(crate) current_log_file_path: String,
+
+    // caching here, becuase the menu building is sync,
+    // but the function to get all languages is async
+    /// Vec of all supported languages: (language_code, language_display_name)
+    pub(crate) all_languages_for_menu: Vec<(String, String)>,
 }
 
 impl AppState {
@@ -36,10 +44,13 @@ impl AppState {
         create_tmp_folder(app.handle()).await?;
         clear_tmp_folder(app.handle()).await?;
 
+        let all_languages_for_menu = get_all_languages(app.handle()).await?;
+
         Ok(Self {
             inner: Arc::new(Mutex::new(InnerAppState::default())),
             startup_timestamp,
             current_log_file_path,
+            all_languages_for_menu,
         })
     }